requirements.txt
Django==2.1.7
django-rest-framework==0.1.0
djangorestframework==3.11.0
mysqlclient==1.4.4
pytz==2019.3
